refactor(web): 按职责拆分 world-detail/game-view/world-select 三组件
- world-detail-view 546→118：抽 detail-primitives(DetailCard/BadgeList/text) + 9 个 Section 到 world-detail-sections
- game-view 528→464：抽 usePanelCollapse + useNavTabActivation hook
- world-select-screen 506→227：抽 WorldCard + WorldListView
- 渲染输出与导出零变更，web 267 测试全过
